Added folder containing parser to give more connections
Handles implicit flows and return statements. In the case of doki doki the game is not fully choice based therefore cannot be fully parsed based of of jump label call menu or even just return statements, this parser handles the return statements and implicit flow of the script execution however this creates a graph with many more edges giving an less accurate representation of the game however still could be helpful to take a look at if you are interested in seeing the implicit flows of the scripts.