Added folder containing parser to give more accurate connections
Does not handle implicit flows however also uses a call stack to help with return statements. Parser produces a graph with same amount of nodes but much less edges nodes without any edges mean that they cannot be reached from a call jump label menu return statement. In the case of doki doki there is alot of different coding styles used and even randomness to get to some of the labels which is why there are no connections for some nodes, the game is not actually fully choice based nor fully renpy based they use some python functions to get to some of the choices in the game.